@@ -33,15 +33,16 @@
     "devDependencies": {
         "@types/chai": "~4.1.2",
         "@types/chai-spies": "~0.0.0",
-        "@types/gl-matrix": "~2.3.2",
-        "@types/mocha": "~2.2.48",
+        "@types/gl-matrix": "~2.4.0",
+        "@types/mocha": "~5.0.0",
         "@types/query-string": "~5.1.0",
         "@types/sinon": "~4.3.0",
         "chai": "~4.1.2",
         "chai-spies": "~1.0.0",
         "concurrently": "~3.5.1",
         "coveralls": "~3.0.0",
-        "gl-matrix": "~2.3.2",
+        "cross-env": "~5.1.4",
+        "gl-matrix": "~2.4.0",
         "glob": "~7.1.2",
         "http-server": "~0.11.1",
         "jsonschema": "~1.2.2",
@@ -50,28 +51,32 @@
         "mocha-multi-reporters": "~1.1.7",
         "nyc": "~11.6.0",
         "pug": "~2.0.1",
-        "query-string": "~5.1.0",
+        "query-string": "~6.0.0",
         "rimraf": "~2.6.2",
         "sinon": "~4.4.6",
         "source-map-support": "~0.5.4",
-        "ts-loader": "~3.5.0",
+        "ts-loader": "~4.1.0",
         "ts-node": "~5.0.1",
         "tslint": "~5.9.1",
         "typedoc": "~0.11.1",
-        "typescript": "~2.7.2",
-        "webpack": "~3.11.0",
-        "webpack-glsl-loader": "~1.0.1",
-        "webpack-sources": "~1.1.0"
+        "typescript": "~2.8.1",
+        "webpack": "^4.3.0",
+        "webpack-bundle-analyzer": "^2.11.1",
+        "webpack-cli": "^2.0.13",
+        "webpack-glsl-loader": "~1.0.1"
     },
     "dependencies": {
         "rxjs": "^5.5.7"
     },
+    "sideEffects": [
+        "./source/polyfill.ts"
+    ],
     "scripts": {
-        "build": "webpack && node pugconfig.js",
+        "build": "webpack-cli --mode development && node pugconfig.js",
         "server": "http-server ./build -a localhost -p 4000",
-        "server-watch": "concurrently \"webpack --watch\" \"node pugconfig.js --watch\" \"npm run-script server\"",
-        "build-dist": "webpack --config webpack.config.dist.js",
-        "build-lib": "webpack --config webpack.config.lib.js",
+        "server-watch": "concurrently \"webpack-cli --mode development --watch\" \"node pugconfig.js --watch\" \"npm run-script server\"",
+        "build-dist": "webpack-cli --mode production --config webpack.config.dist.js",
+        "build-lib": "webpack-cli --mode production --config webpack.config.lib.js",
         "cleanup": "rimraf ./dist ./lib ./build",
         "lint": "tslint --project tsconfig.json --force --format stylish --config tslint.json",
         "test": "nyc mocha",
@@ -79,6 +84,7 @@
         "coverage": "nyc report --reporter=text-lcov | coveralls",
         "doc": "typedoc --hideGenerator --readme README.md --target ES6 --module commonjs --mode file --out dist/doc/ source/",
         "ci": "npm run-script build-dist && npm run-script build-lib && npm run-script lint && npm run-script test",
+        "analyze": "cross-env ANALYZE=true webpack-cli --mode production --config webpack.config.dist.js",
         "prepack": "npm install && concurrently \"npm run build-dist\" \"npm run build-lib\""
     },
     "keywords": [
