@@ -1,6 +1,5 @@
 
-import { Observable } from 'rxjs/Observable';
-import { ReplaySubject } from 'rxjs/ReplaySubject';
+import { Observable, ReplaySubject } from 'rxjs';
 
 import { assert, log, logIf, LogLevel } from './auxiliaries';
 import { clamp } from './gl-matrix-extensions';
