@@ -1,6 +1,5 @@
 
-import { Observable } from 'rxjs/Observable';
-import { Subscription } from 'rxjs/Subscription';
+import { Observable, Subscription } from 'rxjs';
 
 import { assert } from './auxiliaries';
 
