@@ -2,12 +2,12 @@
 import { assert, log, logIf, LogLevel } from './auxiliaries';
 import { GLsizei2 } from './tuples';
 
+import { ChangeLookup } from './changelookup';
 import { Context } from './context';
 import { Framebuffer } from './framebuffer';
 import { Initializable } from './initializable';
 import { NdcFillingTriangle } from './ndcfillingtriangle';
 import { Program } from './program';
-import { ChangeLookup } from './properties';
 import { Shader } from './shader';
 import { Texture2 } from './texture2';
 import { FramePrecisionString, Wizard } from './wizard';
