@@ -1,12 +1,11 @@
 
-export { Canvas } from './canvas';
-
-export { Renderer, Invalidate } from './renderer';
 export { Context } from './context';
-export { ContextMasquerade } from './contextmasquerade';
+export { Canvas } from './canvas';
 export { Controller } from './controller';
+
+export { ContextMasquerade } from './contextmasquerade';
 export { ExtensionsHash } from './extensionshash';
-export { Wizard } from './wizard';
+export { ChangeLookup } from './changelookup';
 export { MouseEventProvider } from './mouseeventprovider';
 export { EventHandler } from './eventhandler';
 
@@ -16,10 +15,12 @@ export { Framebuffer } from './framebuffer';
 export { Geometry } from './geometry';
 export { Program } from './program';
 export { Renderbuffer } from './renderbuffer';
+export { Renderer, Invalidate } from './renderer';
 export { Shader } from './shader';
 export { Texture2 } from './texture2';
 export { TextureCube } from './texturecube';
 export { VertexArray } from './vertexarray';
+export { Wizard } from './wizard';
 
 export { Camera } from './camera';
 export { Color } from './color';
@@ -30,8 +31,8 @@ export { NdcFillingRectangle } from './ndcfillingrectangle';
 export { NdcFillingTriangle } from './ndcfillingtriangle';
 
 export { AntiAliasingKernel } from './antialiasingkernel';
-export { KernelF32, KernelI32, KernelI8, KernelUI32, KernelUI8 } from './kernel';
 export { RandomSquareKernel } from './randomsquarekernel';
+export { KernelF32, KernelI32, KernelI8, KernelUI32, KernelUI8 } from './kernel';
 
 export { AccumulatePass } from './accumulatepass';
 export { BlitPass } from './blitpass';
@@ -43,9 +44,6 @@ export import auxiliaries = root_auxiliaries;
 
 import * as root_gl_matrix_extensions from './gl-matrix-extensions';
 export import gl_matrix_extensions = root_gl_matrix_extensions;
-
-import * as root_properties from './properties';
-export import properties = root_properties;
 
 import * as root_raymath from './raymath';
 export import ray_math = root_raymath;
