@@ -3,9 +3,7 @@ module.exports = require('./webpack.config');
 
 module.exports.entry = {
     '../dist/webgl-operate': ['require.ts', 'polyfill.ts', 'webgl-operate.ts'],
-    '../dist/webgl-operate.min': ['require.ts', 'polyfill.ts', 'webgl-operate.ts'],
     '../dist/webgl-operate.slim': ['require.ts', 'polyfill.ts', 'webgl-operate.slim.ts'],
-    '../dist/webgl-operate.slim.min': ['require.ts', 'polyfill.ts', 'webgl-operate.slim.ts']
 };
 
 module.exports.module.rules[0].use = {
@@ -16,3 +14,8 @@ module.exports.module.rules[0].use = {
         }
     }
 };
+
+if (process.env.ANALYZE) {
+    const analyzer = require('webpack-bundle-analyzer').BundleAnalyzerPlugin;
+    module.exports.plugins.push(new analyzer());
+}
